GM: Resume Required Alert should be triggered off of ECM state (#1045)
* Resume Required Alert should be triggered off of ECM state

On GM cars the ECM goes into cruise state 4 (standstill) and in order
for the ECM to accept cruise inputs, the resume button must be pressed
to drive it to state 1 (active).

Previously rather than testing for if the ECM was in state 4, the wheel
speed sensors were checked that the vehicle was going less than 0.5 m/s.

In some cases this could cause the resume required alert to be displayed
before the car actually came to a stop as far as the ECM was concerned.
If the alert is triggered before state 4 and OP demands acceleration,
its possible for OP to acelerate while the Resume Required Alert is
displayed created a very confusing UX.

* After more review, the problem was actually with the 'if at_full_stop' check was inside an else statement preventing entering the stopped mode if braking is 0

old-commit-hash: 4b12aca9411d2edd178b5972cd0f10c0c02303e8