wider keyboard (#2629)
The keyboard now signals when it wants to open. By subscribing to this signal we can close the sidebar to make the keyboard larger. We also increase the key size to make it easier to type on. 
old-commit-hash: https://github.com/commaai/openpilot/commit/c52e3dfb0887358faee5cb37f0eefc77767fd71d