docs: fix Table of Contents link to "in a car" section (#25225)
* Fix Table of Contents link to "in a car" section

The heading for "Running in a car" changed to "Running on a dedicated device in a car" so the Table of Contents link no longer worked.
This PR updates the link and text in the table.

* Update README.md

Co-authored-by: Shane Smiskol <shane@smiskol.com>
old-commit-hash: https://github.com/commaai/openpilot/commit/a1f4723d1cbdd1fdb4f7d2553e2fb4a07eb7aed6